Remove support for Rails 6.0
