+++ Gemfile
@@ -1,3 +1,3 @@
 # frozen_string_literal: true
 
-eval_gemfile 'gemfiles/rails6.0.gemfile'
+eval_gemfile 'gemfiles/rails6.1.gemfile'
