+++ active_record_host_pool.gemspec
@@ -28,5 +28,5 @@ Gem::Specification.new do |s|
 
   s.required_ruby_version = ">= 2.7.0"
 
-  s.add_runtime_dependency("activerecord", ">= 6.0.0", "< 7.1")
+  s.add_runtime_dependency("activerecord", ">= 6.1.0", "< 7.1")
 end
