--- gemfiles/rails6.0.gemfile
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-source 'https://rubygems.org'
-
-gemspec path: '../'
-
-gem 'activerecord', '~> 6.0.0'
-gem 'mysql2', '>= 0.4.4'
-
-eval_gemfile 'common.rb'
