--- gemfiles/rails6.0.gemfile.lock
@@ -1,84 +1,0 @@
-PATH
-  remote: ..
-  specs:
-    active_record_host_pool (1.2.5)
-      activerecord (>= 6.0.0, < 7.1)
-
-GEM
-  remote: https://rubygems.org/
-  specs:
-    activemodel (6.0.6.1)
-      activesupport (= 6.0.6.1)
-    activerecord (6.0.6.1)
-      activemodel (= 6.0.6.1)
-      activesupport (= 6.0.6.1)
-    activesupport (6.0.6.1)
-      concurrent-ruby (~> 1.0, >= 1.0.2)
-      i18n (>= 0.7, < 2)
-      minitest (~> 5.1)
-      tzinfo (~> 1.1)
-      zeitwerk (~> 2.2, >= 2.2.2)
-    ast (2.4.0)
-    byebug (11.1.1)
-    coderay (1.1.3)
-    concurrent-ruby (1.1.6)
-    i18n (1.8.2)
-      concurrent-ruby (~> 1.0)
-    jaro_winkler (1.5.4)
-    method_source (1.0.0)
-    minitest (5.16.3)
-    minitest-fail-fast (0.1.0)
-      minitest (~> 5)
-    minitest-line (0.6.5)
-      minitest (~> 5.0)
-    minitest-mock_expectations (1.1.3)
-    mysql2 (0.5.5)
-    parallel (1.19.1)
-    parser (2.7.0.2)
-      ast (~> 2.4.0)
-    phenix (1.0.1)
-      activerecord (>= 4.2, < 7.1)
-      bundler
-    pry (0.13.1)
-      coderay (~> 1.1)
-      method_source (~> 1.0)
-    pry-byebug (3.9.0)
-      byebug (~> 11.0)
-      pry (~> 0.13.0)
-    rainbow (3.0.0)
-    rake (13.0.1)
-    rexml (3.2.4)
-    rubocop (0.80.0)
-      jaro_winkler (~> 1.5.1)
-      parallel (~> 1.10)
-      parser (>= 2.7.0.1)
-      rainbow (>= 2.2.2, < 4.0)
-      rexml
-      ruby-progressbar (~> 1.7)
-      unicode-display_width (>= 1.4.0, < 1.7)
-    ruby-progressbar (1.10.1)
-    thread_safe (0.3.6)
-    tzinfo (1.2.6)
-      thread_safe (~> 0.1)
-    unicode-display_width (1.6.1)
-    zeitwerk (2.2.2)
-
-PLATFORMS
-  ruby
-
-DEPENDENCIES
-  active_record_host_pool!
-  activerecord (~> 6.0.0)
-  byebug
-  minitest (>= 5.10.0)
-  minitest-fail-fast
-  minitest-line
-  minitest-mock_expectations (~> 1.1.3)
-  mysql2 (>= 0.4.4)
-  phenix (>= 1.0.1)
-  pry-byebug (~> 3.9)
-  rake (>= 12.0.0)
-  rubocop (~> 0.80.0)
-
-BUNDLED WITH
-   2.4.15
