+++ gemfiles/rails7.0_trilogy.gemfile.lock
@@ -2,7 +2,7 @@ PATH
   remote: ..
   specs:
     active_record_host_pool (1.2.5)
-      activerecord (>= 6.0.0, < 7.1)
+      activerecord (>= 6.1.0, < 7.1)
 
 GEM
   remote: https://rubygems.org/
